+++ src/fs/create.js
@@ -1,6 +1,6 @@
 import { createWriteStream } from 'fs';
 import { isPathExist } from '../utils/isPathExist.js';
-import { color } from '../constants.js';
+import { color } from '../utils/constants.js';
 import { isPathAbsolute } from '../utils/isPathAbsolute.js';
 
 export const create = async (filePath, value) => {
