+++ src/navigation/ls.js
@@ -1,5 +1,5 @@
 import * as fs from 'fs/promises';
-import { color } from '../constants.js';
+import { color } from '../utils/constants.js';
 
 export const listFiles = async (path) => {
   try {
