+++ src/utils/checkInputData.js
