+++ src/zip/compress.js
@@ -1,20 +1,21 @@
 import { createGzip } from 'zlib';
 import { createReadStream, createWriteStream } from 'fs';
 import * as path from 'path';
-import { color } from '../constants.js';
+import { color } from '../utils/constants.js';
+import { isPathAbsolute } from '../utils/isPathAbsolute.js';
 
 export const compressFile = async (filePath, value, pathToDestination) => {
   try {
-    const currentPath = path.join(filePath, value);
+    const currentPath = await isPathAbsolute(filePath, value);
     const compressedFileName = `${path.basename(currentPath)}.gz`;
     const newPath = path.join(pathToDestination, compressedFileName);
     const gzip = createGzip();
     const inp = createReadStream(currentPath);
     const out = createWriteStream(newPath);
     inp.pipe(gzip).pipe(out);
-    console.log(`${color.green}File successfully compressed${color.white}\n`)
+    console.log(`${color.green}File successfully compressed!${color.white}`)
   } catch (error) {
-    console.log(`${color.red}Operation failed${color.white}\n`);
+    console.log(`${color.red}Operation failed!${color.white}`);
   }
 };
